requirements.txt
yfinance==0.2.28
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
ta==0.10.2
plotly==5.17.0
dash==2.14.2
requests==2.31.0
beautifulsoup4==4.12.2
nsepy==0.8
pytz==2023.3
python-dotenv==1.0.0
schedule==1.2.0
loguru==0.7.2
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
sqlite3
threading
datetime
json